Fix type related warning

+++ src/World.cpp
@@ -2554,7 +2554,7 @@ GameObject* World::rayCastClosestOther(const glm::vec3 from, const glm::vec3 dir
     float closestDistance = std::numeric_limits<float>::max();
     size_t closestIndex = std::numeric_limits<size_t>::max();
     if (RayCallback.hasHit()) {
-        for (size_t i = 0; i < RayCallback.m_collisionObjects.size(); ++i) {
+        for (int i = 0; i < RayCallback.m_collisionObjects.size(); ++i) {
             if (RayCallback.m_collisionObjects[i]->getUserPointer() != nullptr) {
                 GameObject* hitObject = static_cast<GameObject *>(RayCallback.m_collisionObjects[i]->getUserPointer());
                 if(hitObject->getWorldObjectID() != ignoreObject->getWorldObjectID()) {
